NEW DataTree widget
