+++ Facades/Elements/UI5DataTree.php
@@ -1,0 +1,292 @@
+<?php
+namespace exface\UI5Facade\Facades\Elements;
+
+use exface\UI5Facade\Facades\Elements\Traits\UI5DataElementTrait;
+use exface\Core\Facades\AbstractAjaxFacade\Elements\JsValueScaleTrait;
+
+/**
+ * 
+ * @method \exface\Core\Widgets\DataTree getWidget()
+ * 
+ * @author Andrej Kabachnik
+ *
+ */
+class UI5DataTree extends UI5DataTable
+{    
+    /**
+     * 
+     * {@inheritDoc}
+     * @see \exface\UI5Facade\Facades\Elements\UI5DataTable::buildJsConstructorForControl()
+     */
+    protected function buildJsConstructorForControl($oControllerJs = 'oController') : string
+    {
+        return $this->buildJsConstructorForTreeTable($oControllerJs);
+    }
+    
+    /**
+     * 
+     * @param string $oControllerJs
+     * @return string
+     */
+    public function buildJsConstructorForTreeTable($oControllerJs = 'oController') : string
+    {
+        $controller = $this->getController();
+        $this->initConfiguratorControl($controller);
+        $widget = $this->getWidget();
+        
+        $selection_mode = $widget->getMultiSelect() ? 'sap.ui.table.SelectionMode.MultiToggle' : 'sap.ui.table.SelectionMode.Single';
+        $selection_behavior = $widget->getMultiSelect() ? 'sap.ui.table.SelectionBehavior.Row' : 'sap.ui.table.SelectionBehavior.RowOnly';
+        
+        if ($widget->getTreeExpandedLevels() !== null) {
+            $numberOfExpandedLevelsJs = "numberOfExpandedLevels: {$widget->getTreeExpandedLevels()},";
+        } else {
+            $numberOfExpandedLevelsJs = "";
+        }
+        
+        return <<<JS
+
+        new sap.ui.table.TreeTable('{$this->getId()}', {
+            {$this->buildJsProperties()}
+            {$this->buildJsPropertyColumnHeight()}
+            rows: {
+                path:'/rows', 
+                parameters: {
+                    arrayNames: [
+                        '_children'
+                    ],
+                    {$numberOfExpandedLevelsJs}
+                }
+            },
+            selectionMode: {$selection_mode},
+	        selectionBehavior: {$selection_behavior},
+    		columns: [
+    			{$this->buildJsColumnsForUiTable()}
+    		],
+            noData: [
+                new sap.m.FlexBox({
+                    height: "100%",
+                    width: "100%",
+                    justifyContent: "Center",
+                    alignItems: "Center",
+                    items: [
+                        new sap.m.Text("{$this->getId()}_noData", {text: "{$this->getWidget()->getEmptyText()}"})
+                    ]
+                })
+            ],
+            toggleOpenState: function(oEvent) {
+                {$this->buildJsOnOpenScript('oEvent')}
+            },
+            firstVisibleRowChanged: function(oEvent) {
+                {$this->buildJsOnCloseScript('oEvent')}
+            }
+        })
+        {$this->buildJsClickHandlers('oController')}
+        {$this->buildJsPseudoEventHandlers()}
+JS;
+    }
+    
+    /**
+     * 
+     * @return string
+     */
+    protected function buildJsPropertyColumnHeight() : string
+    {
+        return '';
+    }
+    
+    /**
+     * 
+     * @param string $oEventJs
+     * @return string
+     */
+    protected function buildJsOnOpenScript(string $oEventJs) : string
+    {
+        return '';
+    }
+    
+    /**
+     * 
+     * @param string $oEventJs
+     * @return string
+     */
+    protected function buildJsOnCloseScript(string $oEventJs) : string
+    {
+        return '';
+    }
+    
+    /**
+     * 
+     * @param string $oModelJs
+     * @return string
+     */
+    protected function buildJsDataLoaderOnLoaded(string $oModelJs = 'oModel') : string
+    {    
+        if (! $this->getWidget()->getTreeParentRelationAlias()) {
+            $treeModeJs = "sap.ui.getCore().byId('{$this->getId()}').setUseFlatMode(true);";
+        } else {
+            $treeModeJs = '';
+        }
+        return parent::buildJsDataLoaderOnLoaded($oModelJs) . <<<JS
+
+                var oDataTree = {$this->buildJsTransformToTree($oModelJs . '.getData()')};
+                {$oModelJs}.setData(oDataTree);
+                {$treeModeJs}
+
+JS;
+    }
+    
+    /**
+     * 
+     * @param string $oControlJs
+     * @return string
+     */
+    protected function buildJsGetRowsAll(string $oControlJs) : string
+    {
+        // NOTE: oTable.getModel().getData() returns only the top level rows, but .getJSON() yields
+        // all. This is why the JSON parsing became neccessary
+        return "({$this->buildJsTransformFromTree("JSON.parse({$oControlJs}.getModel().getJSON()")}).rows || [])";
+    }
+    
+    /**
+     * 
+     * @param string $oDataJs
+     * @return string
+     */
+    protected function buildJsTransformFromTree(string $oDataJs) : string
+    {
+        return <<<JS
+        
+                (function(oDataTree) {
+                    var oDataFlat = $.extend({}, oDataTree);
+                    var fnFlatten = function(aRows) {
+                        var aFlat = [];
+                        aRows.forEach(function(oRow) {
+                            aFlat.push(oRow);
+                            if (Array.isArray(oRow._children) && oRow._children.length > 0) {
+                                aFlat.push(...fnFlatten(oRow._children));
+                            }
+                            delete oRow._children;
+                        });
+                        return aFlat;
+                    };
+                    oDataFlat.rows = fnFlatten(oDataTree.rows || []);                  
+                    return oDataFlat;
+                })($oDataJs)
+                
+JS;
+    }
+    
+    /**
+     * 
+     * @param string $oDataJs
+     * @return string
+     */
+    protected function buildJsTransformToTree(string $oDataJs) : string
+    {
+        $cleanupJs = '';
+        $widget = $this->getWidget();
+        
+        // remove rows without children in oDataTree.rows if $folderFlagAlias is set to 1
+        if (null !== $folderFlagAlias = $widget->getTreeFolderFlagAttributeAlias()) {
+            $cleanupJs = <<<JS
+
+                    for (let i = oDataTree.rows.length - 1; i >= 0; i--) {
+                        (function(oItem, iIndex, aArr) {
+                            if (oItem['{$folderFlagAlias}'] === 1 && oItem['_children'].length === 0) {
+                                aArr.splice(iIndex, 1);
+                            }
+                         })(oDataTree.rows[i], i, oDataTree.rows);
+                    }  
+JS;
+        }
+        
+        return <<<JS
+
+                (function(oDataFlat) {
+                    var oDataTree = $.extend({}, oDataFlat);
+                    var sParentCol = '{$widget->getTreeParentRelationAlias()}';
+
+                    function list_to_tree(list) {
+                      var map = {}, node, roots = [], i;
+                      
+                      for (i = 0; i < list.length; i += 1) {
+                        map[list[i].id] = i; // initialize the map
+                        list[i]._children = []; // initialize the children
+                      }
+                      
+                      for (i = 0; i < list.length; i += 1) {
+                        node = list[i];
+                        // Check, if parent node exists and place the node in its _children array
+                        // Nodes with non-existent parents will be the roots
+                        if (node[sParentCol] !== '' && node[sParentCol] !== null && map[node[sParentCol]] !== undefined) {
+                            list[map[node[sParentCol]]]._children.push(node);
+                        } else {
+                            roots.push(node);
+                        }
+                      }
+                      return roots;
+                    }
+                    if (sParentCol !== '') {
+                        oDataTree.rows = list_to_tree(oDataFlat.rows);
+                    } else {
+                        for (var i = 0; i < oDataTree.rows.length; i++) {
+                            oDataTree.rows[i]._children = [];
+                        }
+                    }
+
+                    $cleanupJs
+                    
+                    return oDataTree;
+                })($oDataJs)
+
+JS;
+    }
+    
+    /**
+     * {@inheritdoc}
+     * @see UI5DataElementTrait::isEditable()
+     */
+    public function isEditable()
+    {
+        return $this->getWidget()->isEditable();
+    }
+    
+    /**
+     * {@inheritdoc}
+     * @see UI5DataElementTrait::hasPaginator()
+     */
+    protected function hasPaginator() : bool
+    {
+        return false;
+    }
+    
+    /**
+     * 
+     * {@inheritDoc}
+     * @see \exface\UI5Facade\Facades\Elements\UI5DataTable::isUiTable()
+     */
+    protected function isUiTable() : bool
+    {
+        return true;
+    }
+    
+    /**
+     * 
+     * {@inheritDoc}
+     * @see \exface\UI5Facade\Facades\Elements\UI5DataTable::isMTable()
+     */
+    protected function isMTable() : bool
+    {
+        return false;
+    }
+    
+    /**
+     * TODO For some reason, transforming model data to tree did not work when there was a dirty-column
+     * 
+     * @return bool
+     */
+    protected function hasDirtyColumn() : bool
+    {
+        return false;
+    }
+}
